requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
